+++ models/ModSchema.js
@@ -3,6 +3,12 @@ const mongoose = require("mongoose")
 let ModSchema = new mongoose.Schema ({
 GuildID: String,
 UserID: String,
+   CurrentMute: {
+        UnmuteAt: Number,
+        Reason: String,
+        Moderator: String,
+        ID: String
+    },
 Punishments: Array
 })